Inspector v2: Export ConvertOptions (#17532)
`ConvertOptions` is a helper function that converts Inspector v1 options
to Inspector v2 options. It is mostly used internally for the
`Inspector` class that exists for back compat. However, it is also used
in Playground. Since I'm changing Playground to consume the Inspector v2
umd bundle, the `ConvertOptions` function needs to be exported and
visible from the umd bundle. This just adds a doc comment and adds it to
the top level exports.