Inspector v2: Fix bug where all SceneExplorer sections can't be collapsed when an item is selected (#17534)
This is to address a bug reported on the forum:
https://forum.babylonjs.com/t/introducing-inspector-v2/60937/105

SceneExplorer selects the currently selected entity when it is first
mounted. The logic for this is wrong and makes it so that if you try to
collapse all top level sections of SceneExplorer, it will trigger
re-selecting and showing the currently selected item, which re-expands
the collapsed section. The intent is to just have this happen when the
component is first mounted. This can just be in a useEffect with no
dependencies, so I'm not sure why I didn't do it this way to start.
🤷🏻‍♂️