Inspector v2: Consume Inspector v2 umd bundle in Sandbox/Playground (#17538)
The initial integration of Inspector v2 into Sandbox/Playground was done
before we had the umd bundle being built and published, and so it was
bundled directly into the host app (Sandbox/Playground). However, to
support versioning (e.g. loading Sandbox/Playground with older
versions), we need to rely on the umd bundle (at least with the current
Sandbox/Playground architecture). This PR changes both apps so that
Inspector v2 is no longer bundled into the host app, but rather the umd
bundle is loaded similar to Inspector v1 and other Babylon dependencies.

Because the umd bundle is being loaded now, when running locally it
loads from dev server, so I also needed to add Inspector v2 to dev
server.

Also because the umd bundle is being loaded now, it means everything is
being loaded, including the back compat for the INSPECTOR global,
BABYLON.Inspector, and Scene.debugLayer. At the same time, we are still
at a point where we need to support switching between v1 and v2. To
handle this, we cache the Inspector v2 module, which already has support
for attaching/detaching these globals, and we dynamically attach/detach
when switching between v1 and v2. Once Inspector v2 is at full parity
with v1, we can simplify this a lot, and simply use v1 when we are
before 8.40.1, and v2 after. I'll make this change within the next
couple of months, before the 9.0 release.

I also updated Playground to include explicit `instantResolve` and
`minVersion` flags for scripts, just like Sandbox.

I also rolled back the webpack configuration for Sandbox and Playground
that allowed us to bundle Inspector v2 (and other dependencies) into the
Sandbox/Playground app bundle (as it is no longer needed).