Make sure Inspector v2 theme matches Playground theme (#17539)
Before the Playground "Editor v2" changes, when the Playground theme
would change, Inspector would get reloaded. This no longer happens,
which on the one hand is nice, but also means the theme of Inspector v2
does not stay in sync with the theme in Playground. For now the easy fix
for this is to just hide and re-show Inspector v2 when the Playground
theme changes. I don't think people do this super often, so losing your
current Inspector state when swapping the theme shouldn't be too
disruptive. What would be better would be to programmatically tell
Inspector v2 to change its theme dynamically, but this would be more
challenging with the current Playground architecture since we'd need
version dependent conditional code (e.g. different Inspector v2 APIs
being available depending on the version). I don't think this approach
would scale at all. We should note that this is another signal for the
value of a newer Playground architecture where we simply rely on dynamic
imports and whole site snapshots for versioning, which would make
handling this trivial.