FrameGraph: optimization of the geometry renderer task (#17543)
The geometry renderer task can now also generate the “beauty” texture,
i.e., render the scene with normal materials.

When applicable, this is an optimization, as you don't need to set up a
geometry renderer for geometry textures and an object render to render
the scene. You can use a single geometry renderer to generate all
textures.

Also in this PR:
* Clear texture task: disable MSAA resolution, as there is no need to
resolve the texture immediately after clearing.
* Clear texture task: check compatibility between target and depth
textures.
* Add a **renderMeshes** property to the object renderer class.