Fix sandbox pan behavior when zoomed close to a large model (#17541)
When zoomed in close to a model, the sandbox code will increase the
panning sensitivity so that small mouse motions don't result in large
unexpected camera movement. However, if you get very close to a model,
the check to see if the inertia should be clamped to zero would start
clamping all slow user movement to zero. Then when the user moved
slightly faster, all of a sudden it wouldn't hit the clamp, and it would
feel like it sped up unexpectedly all of a sudden.

With this fix, we do the same thing that viewer does - scale the speed
parameter as well with radius, so that the clamp gets smaller when close
up to the model.

---------

Co-authored-by: AmoebaChant <kevbrown@microsoft.com>