Smart Filters: add ClampReturnValues debug mode to the optimizer and UI for it in SFE (#17550)
If a shader block relies on the target texture to clamp the output to
0-1, Smart Filters using that block could get different results when run
through the optimizer, since it will combine multiple shader blocks into
a single shader. This adds a debug mode to the optimizer which inserts a
clamp when two shader blocks are combined. If you see different results
when enabling this debug mode, that means you have a block which needs
to clamp its output to ensure it behaves the same when in an optimized
Smart Filter.

---------

Co-authored-by: AmoebaChant <kevbrown@microsoft.com>