GUI: Fix Line layout bug when a connectedControl has a higher zIndex (#17558)
The Line layout code depends on the connectedControl's centerX and
centerY to be up to date, but if the connectedControl had a higher
zIndex, it would do its layout work after the Line control, so the Line
control would be working with stale data during the Control._clip() call
which is part of Control._render(). That stale data meant it would clip
to using the stale location information of the connectedControl, leading
to the bug demonstrated in this playground:
https://playground.babylonjs.com/#5SVAQB#3

With this change, Line will ensure the connectedControl has done its
layout work before it does its own layout math. This ensures the
clipping is calculated correctly regardless of zIndex.

Co-authored-by: AmoebaChant <kevbrown@microsoft.com>