Inspector v2: Rework TexturePreview and use it for Sprite properties (preview sprite texture cell) (#17566)
This PR reworks `TexturePreview` to make the following changes:
- The preview now stretches to fill the available space, with a default
maxWidth of 100% and maxHeight of 384px, both of which can be overridden
at the call site.
- Because the preview can now have an arbitrarily large size, the
texture is now rendered into the canvas at full resolution. This does
mean often we will have a canvas with a larger pixel resolution than
needed, but in the context of Inspector I don't expect this will
actually be a problem in terms of perf or memory use.
- Since `TexturePreview` manages sizing internally now, the `width` and
`height` props were repurposed, and `offsetX` and `offsetY` were added.
All of these props are optional, but they allow a subset of the texture
to be previewed.
- Added a prop to hide the toolbar (channels).
- Used the `TexturePreview` in the sprite properties pane to show a
preview of the sprite cell.

<img width="357" height="304" alt="image"
src="https://github.com/user-attachments/assets/61b037ce-450e-4cc4-9705-45ecdf2b37e5"
/>