Fix legacy sound regression causing only 1 instance to play at a time
The old `Sound` class played multiple overlapping instances when the `play` function was called again before the previous instance finished. When reimplementing the old `Sound` class with the new audio engine, this behavior changed incorrectly to only playing 1 instance, with multiple calls to `play` causing the previous instance to be stopped before playing the new instance.

This change fixes the regression by not setting the `maxInstances` option to `1` so the default of `Infinity` is used instead. This matches the old `Sound` class's behavior.