Fix static sound `stop()` throwing error in some cases (#17572)
When a static sound is created with `autoplay` and `loop` turned on, but
the audio engine is suspended, calling `stop()` on it throws a WebAudio
error because the underlying source node is not playing.

This change fixes the issue by making sure the sound has actually
started playing before calling `stop` on the underlying WebAudio source
node.