Inspector v2: Texture Editor (#17578)
This PR adds a Texture Editor implementation for Inspector v2.
-❗This is not a complete re-write. The React independent code (like
CanvasManager) is just copied from v1 with only minor changes for build
rules like TS strict mode used in v2.
- For all the React code, the first pass was created by GitHub Copilot
(Claude Opus 4.5) as a port of the v1 React components.
- From there, I cleaned up the React code and fixed some issues.
- And also in v1, there was already an abstraction for Texture Editor
"tools" (paint, fill, contrast, etc.) with somewhat of an extensibility
mechanism (though I don't think it was reachable through the Inspector
v1 API). Since this abstraction was already in place, I used it and made
Texture Editor tools another extensibility point in the default services
for Inspector v2. This means extensions can add new Texture Editor
tools.

Other notable changes:
- We already had code for handling Fluent in child windows for undocking
Inspector v2 side panes. We need the same thing for Texture Editor and
various other features across Inspector v2 and our other tools, so I
factored out a `ChildWindow` component in shared-ui-controls that does
all the setup to make Fluent work in a child window. It uses a react
imperative ref to get a handle for opening/closing the child window, and
has a callback prop to be notified of open state change.
- Added a react imperative ref to the existing `TexturePreview`
component so we can force a refresh of the texture data when needed
(e.g. when `TextureEditor` modifies the texture).
- Fixed an issue where in some cases, we were applying styles through
classes to FluentProvider for layout, which by default is applied to
portals, which can break things. We need to either use inline styles
only for layout related styling, or prevent styles from flowing to
portals (at least by default). I went with the latter as it seems like
the best option to prevent us from hitting this issue in the future.
- Updated ShellService to use the new `ChildWindow` component.
- Factored out `ApplyChannelsToTextureDataAsync` from `TexturePreview`
into a helper as it is used by `TextureEditor` as well.
- Updated `Popover` and `ColorPicker` to use `forwardRef` so they can be
used with components that need a ref to the html element (like
`Tooltips` and `Portals`).

<img width="936" height="710" alt="image"
src="https://github.com/user-attachments/assets/d8b432d8-f506-46b4-8d0d-b4e0e83baf20"
/>