[GlowLayer] Add an option to set all meshes excluded from the glow layer (#17581)
This PR adds a new boolean/option for the glow layer, called
`excludeByDefault`. It reverses the default state how meshes are
included in the glow layer. By default, when glow layer is initialized,
all the meshes are candidates for the glow layer. If the
`excludeByDefault` is uses, one needs to manually add meshes to be
included to the glow layer. If this option is set to `true`, the
excluded meshes array does nothing and there is a comment added to
indicate that to the js docs.

The forum thread about the issue:

https://forum.babylonjs.com/t/problem-excluding-everything-out-of-glow-layer-by-default/61772