EXRCubeTexture: Fix spherical harmonics (#17586)
See
https://forum.babylonjs.com/t/exrcubetexture-with-no-polynomials/61867

When creating an EXR cube texture (but it's probably the same with
`HDRCubeTexture`), the code that sets default spherical polynomials
arrived too late in the loading process, and
`PBRBaseMaterial.isReadyForSubMesh` had time to execute, which triggered
the calculation of the polynomials. The fix ensures that the default
polynomials are defined as soon as the internal texture is created.

Note that I had to update a number of visualization tests, because they
passed `false` for the **generateHarmonics** parameter of the
`HDRCubeTexture` constructor.

Supersedes #17584.