Meshes: Fix wrong normals after calling updateFacetData (#17593)
See
https://forum.babylonjs.com/t/mesh-updatefacetdata-breaks-smooth-shading-after-gltf-glb-export-faceted-look-on-rounded-meshes/61944

The problem is that calling `updateFacetData` updates the normal array
on the CPU side, but not the GPU buffer. So the mesh looks fine in the
Playground, but after exporting/importing it, it no longer does, because
when exporting, we use the CPU-side array.

I don't think `updateFacetData` was designed to modify a mesh's normal
array (I don't see why we would do that?), so I simply cloned the array
before calling `VertexData.ComputeNormals` (which is the function that
updates the normals).